services test (works) + rest

+++ src/main/java/artezio/vkolodynsky/repository/SessionRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SessionRepository extends CrudRepository<Session, Long> {
-
+public interface SessionRepository extends CrudRepository<Session, Integer> {
 }
