+++ src/main/java/artezio/vkolodynsky/service/SessionService.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface SessionService {
     List<Session> findAll();
     Session save(Session value);
-    void deleteByID(long id);
-    Session findByID(long id);
+    void deleteByID(int id);
+    Session findByID(int id);
 }
